pyproject.toml
[project]
name = "video-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromium>=0.0.0",
    "curl-cffi>=0.14.0",
    "ffmpeg>=1.4",
    "nodriver>=0.48.1",
    "pillow>=12.1.0",
    "playwright>=1.57.0",
    "python-vlc>=3.0.21203",
    "requests>=2.32.5",
    "yt-dlp>=2025.12.8",
]
